feat: refactor DetailPage component logic into custom useDetailPage hook
