+++ src/pages/Detailpage.tsx
@@ -1,45 +1,15 @@
 import { Breadcrumb, Related, Loading, ProductInfo } from "../utils";
-import { useParams } from "react-router-dom";
-import { useGetItemsQuery, useGetItemBySlugQuery } from "../store/pokemartApi";
 import ProductNotFound from "../components/detailpage/ProductNotFound";
-import { useMemo } from "react";
-import { getRandomItems } from "../utils/arrayUtils";
+import { useDetailPage } from "../hooks/useDetailPage"; // Custom hook for logic
 
 const DetailPage = () => {
-  const { slug } = useParams();
-  const { data: Item, isLoading: loading } = useGetItemBySlugQuery(slug ?? "", {
-    skip: !slug,
-  });
+  const { Item, loading, slug, randomProducts } = useDetailPage();
 
-  const { data: countData } = useGetItemsQuery({
-    page: 1,
-    limit: 1,
-  });
-
-  const { data: allProductsData } = useGetItemsQuery(
-    {
-      page: 1,
-      limit: countData?.info.count ?? 1,
-    },
-    {
-      skip: !countData?.info.count,
-    },
-  );
-
-  const randomProducts = useMemo(() => {
-    if (!allProductsData?.items) return [];
-    return getRandomItems(allProductsData.items, 5);
-  }, [allProductsData?.items]);
-
-  if (loading) {
-    return <Loading />;
-  }
-
-  if (!Item) {
+  if (loading) return <Loading />;
+  if (!Item)
     return (
       <ProductNotFound name={slug ?? ""} availableProducts={randomProducts} />
     );
-  }
 
   return (
     <>
